Backend/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
python-dotenv>=1.0.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=1.1.0
openpyxl>=3.1.2

# AI/ML for RAG
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
sentence-transformers>=2.2.2
chromadb>=0.4.0
tiktoken>=0.5.1

# Async Processing
celery>=5.3.4
redis>=5.0.1
django-celery-beat>=2.5.0
django-celery-results>=2.5.1

# Testing
pytest>=7.4.3
pytest-django>=4.7.0
